requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
click==8.0.3
colorama==0.4.4
cvxopt==1.2.7
cvxpy==1.1.18
cycler==0.11.0
ecos==2.0.10
fancyimpute==0.7.0
Flask==2.0.2
fonttools==4.29.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
knnimpute==0.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
nose==1.3.7
numpy==1.22.2
opencv-python==4.5.5.62
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.0
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
pytz==2021.3
qdldl==0.1.5.post0
scikit-learn==1.0.2
scipy==1.7.3
scs==3.1.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tomli==2.0.0
Werkzeug==2.0.2
